Rename search module to tree
