+++ paramtools/tree.py
